ai-server/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.30.6

# LLMs
openai==1.45.0
typing-extensions==4.12.2

# Env management
python-dotenv==1.0.1

# Database (MongoDB, async)
motor==3.6.0

# Image handling
Pillow==10.4.0
requests==2.32.3

# Cloudinary
cloudinary==1.41.0

# Google Gemini (for cleanup image analysis — replaces google-cloud-vision)
google-generativeai>=0.7.0

# Utils
annotated-types==0.7.0
